Users now have a professional-grade command line interface with:   - 35+ built-in commands across 7 categories   - Advanced system monitoring like a Linux server   - Enterprise network management with WiFi optimization   - Comprehensive power management with battery analytics   - System snapshot capabilities for performance tracking   - Intelligent recommendations for optimization   - Professional-grade diagnostics and troubleshooting tools

+++ app/src/main/java/com/win11launcher/data/database/DatabaseCleanupManager.kt
@@ -33,7 +33,7 @@ class DatabaseCleanupManager @Inject constructor(
                 } catch (e: Exception) {
                     // Log error but don't crash
                 }
-                delay(TimeUnit.HOURS.toMillis(CLEANUP_INTERVAL_HOURS))
+                delay(TimeUnit.HOURS.toMillis(CLEANUP_INTERVAL_HOURS.toLong()))
             }
         }
     }
